% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMonitor.R
\name{callMonitor}
\alias{callMonitor}
\title{Helper to call certain step function of a monitor.}
\usage{
callMonitor(monitor, step, envir = parent.frame())
}
\arguments{
\item{monitor}{[\code{CMAES_monitor}]\cr
Monitor.}

\item{step}{[\code{character(1)}]\cr
One of before, step, after.}

\item{envir}{[\code{environment}]\cr
The environment to pass.}
}
\description{
This funtions serves to call a specific monitor step.
}

