% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeStoppingCondition.R
\name{getDefaultStoppingConditions}
\alias{getDefaultStoppingConditions}
\title{Return list of default stopping conditions.}
\usage{
getDefaultStoppingConditions()
}
\value{
[\code{list}]
}
\description{
Return list of default stopping conditions.
}

