% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMonitor.R
\name{makeMonitor}
\alias{makeMonitor}
\title{Factory method for monitor objects.}
\usage{
makeMonitor(before = NULL, step = NULL, after = NULL, ...)
}
\arguments{
\item{before}{[\code{function}]\cr
Function called one time after initialization of the EA.}

\item{step}{[\code{function}]\cr
Function applied after each iteration of the algorithm.}

\item{after}{[\code{function}]\cr
Function applied after the EA terminated.}

\item{...}{[\code{any}]\cr
Not used.}
}
\value{
[\code{cma_monitor}]
  Monitor object.
}
\description{
Monitors can be pluged in the main \code{\link{cmaes}} function.
They have full access to the environment of the optimization routine and can
be used to write/log/visualize relevant data in each iteration.
}
\seealso{
\code{\link{makeSimpleMonitor}}, \code{\link{makeVisualizingMonitor}}
}

