% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeSimpleMonitor}
\alias{makeSimpleMonitor}
\title{Generator for simple monitor.}
\usage{
makeSimpleMonitor()
}
\value{
[\code{cma_monitor}]
}
\description{
The simple monitor prints the iteration, current best parameter values and best fitness
to the standard output.
}

