% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMonitor.R
\name{makeVisualizingMonitor}
\alias{makeVisualizingMonitor}
\title{Generator for visualizing monitor.}
\usage{
makeVisualizingMonitor(show.last = FALSE, show.distribution = TRUE)
}
\arguments{
\item{show.last}{[\code{logical(1)}]\cr
Should the last population be visualized as well?
Default is \code{FALSE}.}

\item{show.distribution}{[\code{logical(1)}]\cr
Should an ellipsis of the normal distribution be plotted?
Default is \code{TRUE}.}
}
\value{
[\code{cma_monitor}]
}
\description{
This generator visualizes the optimization process for two-dimensional functions
by means of \pkg{ggplot2}.
}
\details{
The plot contains points representing the current population, the center
of mass or mean value of the population respectively. Optionally an ellipsis
represneting the normal distribution of the points can be depicted.
}

