% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runCMAES}
\alias{runCMAES}
\title{Covariance-Matrix-Adaption}
\usage{
runCMAES(objective.fun, start.point, population.size = NULL, sigma,
  max.iter = 10L, monitor = makeSimpleMonitor())
}
\arguments{
\item{objective.fun}{[\code{\link[otf]{otf_function}}]\cr
Numerical objective function of type \code{\link[otf]{otf_function}}. The function
must expect a list of numerical values and return a scaler numerical value.}

\item{start.point}{[\code{numeric}]\cr
Initial solution vector.}

\item{population.size}{[\code{integer(1)}]\cr
Population size.}

\item{sigma}{[\code{numeric(1)}]\cr
Initial step-size, i. e., standard deviation in each coordinate direction.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of sequential iterations.}

\item{monitor}{[\code{cma_monitor}]\cr
Monitoring object.}
}
\value{
[\code{CMAES_result}] Result object.
}
\description{
Performs non-linear, non-convex optimization by means of the Covariance
Matrix Adaption Evolutionary Strategy by Hansen.
}

