% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnCondCov}
\alias{stopOnCondCov}
\title{Stopping condition: high condition number.}
\usage{
stopOnCondCov(tol = 1e+14)
}
\arguments{
\item{tol}{[\code{numeric(1)}]\cr
Tolerance value.
Default is \code{1e14}.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if condition number of covariance matrix exceeds
tolerance value.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnIndefCovMat}},
  \code{\link{stopOnMaxIters}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnNoEffectCoord}},
  \code{\link{stopOnOptParam}},
  \code{\link{stopOnOptValue}},
  \code{\link{stopOnTimeBudget}}
}

