% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnIndefCovMat}
\alias{stopOnIndefCovMat}
\title{Stopping condition: stop if covariance matrix is not positive definite anymore.}
\usage{
stopOnIndefCovMat()
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if covariance matrix is not positive definite anymore.
}

