% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnMaxEvals}
\alias{stopOnMaxEvals}
\title{Stopping condition: maximal funtion evaluations.}
\usage{
stopOnMaxEvals(max.evals)
}
\arguments{
\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of allowed function evaluations.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if maximal number of function evaluations is reached.
}

