% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnNoEffectCoord}
\alias{stopOnNoEffectCoord}
\title{Stopping condition: stop if addition of 0.2 standard deviations in any
coordinate does not change mean value.}
\usage{
stopOnNoEffectCoord()
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if addition of 0.2 standard deviations in any
coordinate does not change mean value.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}};
  \code{\link{stopOnIndefCovMat}};
  \code{\link{stopOnMaxIters}};
  \code{\link{stopOnNoEffectAxis}};
  \code{\link{stopOnOptParam}};
  \code{\link{stopOnOptValue}};
  \code{\link{stopOnTimeBudget}}
}

