% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnOptParam}
\alias{stopOnOptParam}
\title{Stopping condition: optimal params.}
\usage{
stopOnOptParam(opt.param, tol = 1e-08)
}
\arguments{
\item{opt.param}{[\code{numeric}]\cr
Known optimal parameter settings.}

\item{tol}{[\code{numeric(1)}]\cr
Tolerance value.
Default is \eqn{1e^{-8}}.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if euclidean distance of parameter is below
some tolerance value.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}},
  \code{\link{stopOnIndefCovMat}},
  \code{\link{stopOnMaxIters}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnNoEffectCoord}},
  \code{\link{stopOnOptValue}},
  \code{\link{stopOnTimeBudget}}
}

