% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnOptValue}
\alias{stopOnOptValue}
\title{Stopping condition: stop if best solution is close to optimal objective value.}
\usage{
stopOnOptValue(opt.value, tol = 1e-08)
}
\arguments{
\item{opt.value}{[\code{numeric(1)}]\cr
Known optimal objective function value.}

\item{tol}{[\code{numeric(1)}]\cr
Tolerance value.
Default is \eqn{1e^{-8}}.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if best solution is close to optimal objective value.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}};
  \code{\link{stopOnIndefCovMat}};
  \code{\link{stopOnMaxIters}};
  \code{\link{stopOnNoEffectAxis}};
  \code{\link{stopOnNoEffectCoord}};
  \code{\link{stopOnOptParam}};
  \code{\link{stopOnTimeBudget}}
}

