% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnTimeBudget}
\alias{stopOnTimeBudget}
\title{Stopping condition: stop if maximal running time budget is reached.}
\usage{
stopOnTimeBudget(budget)
}
\arguments{
\item{budget}{[\code{integer(1)}]\cr
Time budget in seconds.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if maximal running time budget is reached.
}

