% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnTolX}
\alias{stopOnTolX}
\title{Stopping condition: stop if the standard deviation falls below a tolerance value
in all coordinates?}
\usage{
stopOnTolX(tol = 10^(-12))
}
\arguments{
\item{tol}{[\code{integer(1)}]\cr
Tolerance value.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if the standard deviation falls below a tolerance value
in all coordinates?
}

