% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStoppingCondition.R
\name{getDefaultStoppingConditions}
\alias{getDefaultStoppingConditions}
\title{Return list of default stopping conditions.}
\usage{
getDefaultStoppingConditions()
}
\value{
[\code{list}]
}
\description{
Default stopping conditions which are active in the reference
implementation by Nico Hansen in Python.
}

