% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeMonitor.R
\name{makeVisualizingMonitor}
\alias{makeVisualizingMonitor}
\title{Generator for visualizing monitor.}
\usage{
makeVisualizingMonitor()
}
\value{
[\code{cma_monitor}]
}
\description{
This generator visualizes the optimization process for two-dimensional functions
by means of \pkg{ggplot2}.
}

