% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runCMAES.R
\name{runCMAES}
\alias{runCMAES}
\title{Covariance-Matrix-Adaption}
\usage{
runCMAES(objective.fun, start.point = NULL, population.size = NULL, sigma,
  max.iter = 10L, max.evals = Inf, max.time = Inf,
  monitor = makeSimpleMonitor())
}
\arguments{
\item{objective.fun}{[\code{smoof_function}]\cr
Numerical objective function of type \code{smoof_function}. The function
must expect a list of numerical values and return a scaler numerical value.}

\item{start.point}{[\code{numeric}]\cr
Initial solution vector. If \code{NULL}, one is generated randomly within the
box constraints offered by the paramter set of the objective function.}

\item{population.size}{[\code{integer(1)}]\cr
Population size.}

\item{sigma}{[\code{numeric(1)}]\cr
Initial step-size, i. e., standard deviation in each coordinate direction.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of sequential iterations.
Default is 10.}

\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of function evaluations.
Default is \code{Inf}.}

\item{max.time}{[\code{integer(1)}]\cr
Maximal time budget in seconds.
Default is \code{Inf}.}

\item{monitor}{[\code{cma_monitor}]\cr
Monitoring object.}
}
\value{
[\code{CMAES_result}] Result object.
}
\description{
Performs non-linear, non-convex optimization by means of the Covariance
Matrix Adaption Evolutionary Strategy by Hansen.
}

