% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnMaxEvals}
\alias{stopOnMaxEvals}
\title{Stopping condition: stop if maximal number of function evaluations is reached.}
\usage{
stopOnMaxEvals(max.evals)
}
\arguments{
\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of allowed function evaluations.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if maximal number of function evaluations is reached.
}

