% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnMaxIters}
\alias{stopOnMaxIters}
\title{Stopping condition: maximal iterations.}
\usage{
stopOnMaxIters(max.iter = 100L)
}
\arguments{
\item{max.iter}{[integer(1)]\cr
Maximal number of iterations.
Default is \code{100}.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop on maximal number of iterations.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}},
  \code{\link{stopOnIndefCovMat}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnNoEffectCoord}},
  \code{\link{stopOnOptParam}},
  \code{\link{stopOnOptValue}},
  \code{\link{stopOnTimeBudget}}
}

