% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnNoEffectAxis}
\alias{stopOnNoEffectAxis}
\title{Stopping condition: stop if addition of 0.1 * sigma in a principal axis
direction change mean value?}
\usage{
stopOnNoEffectAxis()
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stopping condition: stop if addition of 0.1 * sigma in a principal axis
direction change mean value?
}
\seealso{
Other stopping.conditions: \code{\link{stopOnIndefCovMat}};
  \code{\link{stopOnMaxIters}};
  \code{\link{stopOnNoEffectCoord}};
  \code{\link{stopOnOptParam}};
  \code{\link{stopOnOptValue}};
  \code{\link{stopOnTimeBudget}}
}

