% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnNoEffectCoord}
\alias{stopOnNoEffectCoord}
\title{Stopping condition: standard deviation in coordinates.}
\usage{
stopOnNoEffectCoord()
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if addition of 0.2 * standard deviations in any
coordinate does not change mean value.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}},
  \code{\link{stopOnMaxIters}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnOptParam}},
  \code{\link{stopOnOptValue}},
  \code{\link{stopOnTimeBudget}}
}

