% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnTimeBudget}
\alias{stopOnTimeBudget}
\title{Stopping condition: maximal time.}
\usage{
stopOnTimeBudget(budget)
}
\arguments{
\item{budget}{[\code{integer(1)}]\cr
Time budget in seconds.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if maximal running time budget is reached.
}
\seealso{
Other stopping.conditions: \code{\link{stopOnCondCov}},
  \code{\link{stopOnIndefCovMat}},
  \code{\link{stopOnMaxIters}},
  \code{\link{stopOnNoEffectAxis}},
  \code{\link{stopOnNoEffectCoord}},
  \code{\link{stopOnOptParam}},
  \code{\link{stopOnOptValue}}
}

