% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoppingConditions.R
\name{stopOnTolX}
\alias{stopOnTolX}
\title{Stopping condition: low standard deviation.}
\usage{
stopOnTolX(tol = 1e-12)
}
\arguments{
\item{tol}{[\code{integer(1)}]\cr
Tolerance value.}
}
\value{
[\code{cma_stopping_condition}]
}
\description{
Stop if the standard deviation falls below a tolerance value
in all coordinates?
}

